Updating build format
This is an overhaul to the demo pipeline narrative. I wanted to do this to show that config data changes are a parallel part of the application build process (in addition to code). These pipeline steps better illustrate this fact and also give change management an explicit step to call out during demo.